Created hashcode and the equals methods.
