--- src/main/Monster.java
@@ -1,37 +1,0 @@
-public abstract class Monster {
-    private Integer hp;
-    private Integer xp;
-    private Integer maxHP;
-    private HashMap<String, Integer> items;
-
-    public Monster(Integer maxHP, Integer xp, HashMap<String, Integer> items) {
-        this.maxHP = maxHP;
-        hp = this.maxHP;
-        this.xp = xp;
-        this.items = items;
-    }
-
-    public Integer getXp() {
-        return xp;
-    }
-
-    public Integer getMaxHP() {
-        return maxHP;
-    }
-
-    public Integer getHp() {
-        return hp;
-    }
-
-    public void setHp(Integer hp) {
-        this.hp = hp;
-    }
-
-    public HashMap<String, Integer> getItems() {
-        return items;
-    }
-
-    public void setItems(HashMap<String, Integer> items) {
-        this.items = items;
-    }
-}
